split mteb_utils
